Do not optimize away always-true ResetIf with hit count
An expression such as `repeated(2, never(1 == 1))` can be used to reset an
achievement every other frame. This expression is converted into a requirement
`1 == 1` with type `ResetIf` and an `HitCount` of 2.

The optimizer was optimizing the whole requirement away, because `1 == 1` is
always true. Even though that's correct, the removal of the `ResetIf` completely
changes the achievement's logic. Fixed by not removing always-true requirements
if their type is either `ResetIf` or `PauseIf`.

Another issue was that the optimizer also considered `repeated(2, 1 == 1)` to
be always true and replaced it with `1 == 1`, removing the hit count that
could be relevant for later (such as when wrapped in a `never` call). Fixed by
setting the hit count of the injected `1 == 1` requirement to be the maximum hit
count of the replaced always-true requirements.