Support for fallback value in rich presence lookups
As specified in the Rich Presence docs: If a lookup doesn't contain a mapping
for a value, it will result in a blank (no space) value. You can change this
behavior by adding a single line to the lookup mapping '*' to the desired
fallback value.

Added support for that in RATools. The lookup map simply has to include a
"*" string constant key.