correctly dump tickets with empty and non-breakable lines
If a ticket's text contained a large line without a space, RATools
would fail to dump it and crash whilst trying to break it into smaller
lines. Happened with a URL to a proof image.

Also noticed that empty lines were not dumped correctly: the newline
would be missing and the following line would be double-commented:

```
//               //               next line
```